Changed sql data types to utilize floats

+++ database/models.py
@@ -1,5 +1,5 @@
 import datetime
-from sqlalchemy import Numeric, DateTime, func
+from sqlalchemy import DateTime, func, Float
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
 class Base(DeclarativeBase):
@@ -10,9 +10,9 @@ class SensorReads(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True)
 
-    temp: Mapped[float] = mapped_column(Numeric)
-    humidity: Mapped[float] = mapped_column(Numeric)
-    pressure: Mapped[float] = mapped_column(Numeric)
+    temp: Mapped[float] = mapped_column(Float)
+    humidity: Mapped[float] = mapped_column(Float)
+    pressure: Mapped[float] = mapped_column(Float)
 
     timestamp: Mapped[datetime.datetime] = mapped_column(
         DateTime(timezone=True),
